3-Multi-Agent-System/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain-google-genai==1.0.0
langgraph==0.0.39
python-dotenv==1.0.0
pydantic==2.5.0
